[INTERNAL] Refactoring and documentation of api (#244)
