+++ .github/workflows/sphinx-build.yml
@@ -30,12 +30,15 @@ jobs:
             SPHINXOPTS: -W  --keep-going
         run: |
           make html
-      - name: Check general doc
+      - name: Check general doc coverage
         run: |
-          docstr-coverage . -e .env -F 40
-      - name: Check models doc
+          docstr-coverage . -e .env -F 59
+      - name: Check models doc coverage
         run: |
           docstr-coverage collectives/models
-      - name: Check utils doc
+      - name: Check utils doc coverage
         run: |
           docstr-coverage collectives/utils
+      - name: Check api doc coverage
+        run: |
+          docstr-coverage collectives/api
