+++ doc/source/api.rst
@@ -2,4 +2,35 @@ Module ``collectives.api``
 =======================================
 
 .. automodule:: collectives.api
-      :members:
+        :members:
+
+
+Module ``collectives.api.common``
+---------------------------------------
+.. automodule:: collectives.api.common
+        :members:
+
+
+Module ``collectives.api.admin``
+---------------------------------------
+.. automodule:: collectives.api.admin
+        :members:
+
+
+Module ``collectives.api.autocomplete_user``
+--------------------------------------------------
+.. automodule:: collectives.api.autocomplete_user
+        :members:
+
+
+Module ``collectives.api.event``
+---------------------------------------
+.. automodule:: collectives.api.event
+        :members:
+
+
+
+Module ``collectives.api.userevent``
+---------------------------------------
+.. automodule:: collectives.api.userevent
+        :members:
